+++ Lab_4/Shape/Point.h
@@ -3,11 +3,11 @@ class CPoint final
 {
 public:
     CPoint() = default;
-    CPoint(double x, double y);
+    CPoint(float x, float y);
     ~CPoint() = default;
     std::string ToString() const;
-    double x = 0.0; 
-    double y = 0.0;
+    float x = 0.0; 
+    float y = 0.0;
 };
 
 bool operator >> (std::istream & in, CPoint & point);
